+++ MagicalLifeAPI/Sound/MusicPlayer.cs
@@ -30,6 +30,8 @@ namespace MagicalLifeAPI.Sound
 
             MusicPlayer.SONG_MAIN_MENU = Instance.LoadSong("MainMenu1");
             //Instance.Play(SONG_MAIN_MENU);
+            FMODUtil.Init();
+            FMODUtil.Test();
         }
 
         public void Unload()