+++ MagicalLifeGUIWindows/GUI/MainMenu/Buttons/NewGameButton.cs
@@ -1,4 +1,5 @@
 ﻿using MagicalLifeAPI.Networking.Client;
+using MagicalLifeAPI.Sound;
 using MagicalLifeGUIWindows.GUI.Reusable;
 using Microsoft.Xna.Framework;
 using MonoGame.Extended.Input.InputListeners;
@@ -13,6 +14,7 @@ namespace MagicalLifeGUIWindows.GUI.MainMenu.Buttons
 
         public override void Click(MouseEventArgs e)
         {
+            FMODUtil.RaiseEvent(EffectsTable.UIClick);
             New_World_Menu.NewWorldMenu.Initialize();
         }