+++ MagicalLifeGUIWindows/Load/Initializer.cs
@@ -21,7 +21,6 @@ namespace MagicalLifeGUIWindows.Load
             InputHandlers.Initialize();
             SettingsManager.Initialize();
             MusicPlayer.Init();
-            EffectPlayer.Init();
         }
     }
 }