+++ CHANGELOG.md
@@ -10,6 +10,7 @@ SneakyTactician  <SneakyTactician@outlook.com>
 * Added a grass tile
 * Added the stone chunk item
 * Mining jobs are now possible
+* Added internal support for sfx/sound events
 
 #### Server
 * Seperated server loading and server starting into two seperate commands
@@ -18,6 +19,7 @@ SneakyTactician  <SneakyTactician@outlook.com>
 #### GUI
 * Grass textures now connect to form smoother patterns
 * Now showing logos for MonoGame and FMOD
+* Buttons now play a clicking sound when clicked upon
 * Player can now mark stone to be mined
 
 #### Bugs